Merged PR 5: PBI-277: Pipeline initial commit
This PR is the initial commit for the cdisc standards pipeline. It is meant to be a pipeline for converting wiki documents into json documents to be stored in cdisc library blob storage.

**Steps to Test**
In order to test the pipeline you'll need to setup a virtual environment and supply a config/environment variables for metadata information. Setup steps can be found in the readme. Below are the config values I used to test each product type.

sendig: 113587958
sdtm: summary 113598689, classMetadata = 103756571, datasetMetadata = 103756571
cdash: 113597660
cdashig: 113592763
adamig: 113585917

In all cases it should not be necessary to specify a `variables` config value but if necessary the same document id used for the summary can be used for the variables.

Once the environment is set up and the config file is created simply run the following command from the project root to run the pipeline

```
python .\parse_document.py -u '<confluence_username>' -p '<confluence_password>' -a '<api_key>' -l '<log_level>' -c  '<path to config file>' -i
```

Please let me know if you need any credentials or have any questions

Related work items: #277