Merged PR 80: PIPELINE V2
This PR is a rewrite of the standard products pipeline, with the goal of making the codebase easier to debug and extend. This PR also fixes issues with the CDASH mapping target logic.

The main change in this PR is that logic pertaining to certain top level objects: (scenarios, domains, datasets, classes, variables...) is now extracted into an appropriate class for that object.

Steps to Test:

Run the tests with the following command from the project root: `pytest`